requirements.txt
django==3.1.7
petl==1.7.4
requests==2.25.1
python-dateutil==2.8.1